build.sh: Update Target check and remove bytecode creation
Currently in build.sh the target function has techpack command which
enables the tech pack command and while running QIIFA through build.sh,
QIIFA is generating bytcode(pyc) files.

This change is update the target function to remove the techpack command
and to add the bytcode flag in python to remove the generation of
bytecode(pyc) files will not be generated.

Change-Id: I64c5bf2185640dcfd171b532b2aa76a0a0b1eb6d