core-utils: Remove IMinkTransact from matrix
Server and client communicates over IMinkTrasanct by creating a binder object
of IMinkTransact. Client passes this binder object during connect
and server returns its IMinkTransact binder object and this is how this interface is later used.
There is no service which does addservice on IMinkTransact hence this entry is not needed in
manifest file.

Change-Id: I7d74b3236532bc8ab032a3f54ed58722e1a41003